requirements.txt
mautrix>=0.15.0,<0.16
aiohttp>=3,<4
yarl>=1,<2
SQLAlchemy>=1,<1.4
alembic>=1,<2
commonmark>=0.9,<1
ruamel.yaml>=0.15.35,<0.18
attrs>=18.1.0
bcrypt>=3,<4
packaging>=10

click>=7,<8
colorama>=0.4,<0.5
questionary>=1,<2
jinja2>=2,<4
